simplified store config
