--- src/store/configureStore.dev.js
@@ -1,35 +1,0 @@
-import { createStore, applyMiddleware, compose } from 'redux';
-import createLogger from 'redux-logger';
-import createSagaMiddleware from 'redux-saga';
-
-import reducer from '../reducers';
-import saga from '../sagas';
-
-export default function configureStore() {
-  const sagaMiddleware = createSagaMiddleware();
-
-  const store = createStore(
-    reducer,
-    compose(
-      applyMiddleware(
-        sagaMiddleware,
-          createLogger()
-      ),
-      window.devToolsExtension ? // eslint-disable-line no-undef
-        window.devToolsExtension() : // eslint-disable-line no-undef
-        value => value
-    )
-  );
-
-  sagaMiddleware.run(saga);
-
-  if (module.hot) {
-    module.hot.accept('../reducers', () => {
-      const nextReducer = require('../reducers').default; // eslint-disable-line global-require
-
-      store.replaceReducer(nextReducer);
-    });
-  }
-
-  return store;
-}
